Move mail config to config file
